requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
altair==4.2.2
analytics-python==1.4.post1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.1.0
av==12.3.0
backoff==1.10.0
base58==2.1.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
better-profanity==0.7.0
bidict==0.23.1
blinker==1.8.2
bs4==0.0.2
cachetools==5.4.0
certifi==2024.6.2
cffi==1.17.0
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
comm==0.2.2
compressed_rtf==1.0.6
contourpy==1.2.1
cryptography==43.0.1
cycler==0.12.1
debugpy==1.8.6
decorator==4.4.2
Deprecated==1.2.14
discord==2.3.2
discord.py==2.3.2
dnspython==2.6.1
easygui==0.98.3
ebcdic==1.1.1
eml-parser==1.17.5
entrypoints==0.4
et-xmlfile==1.1.0
eventlet==0.36.1
exceptiongroup==1.2.2
executing==2.1.0
extract-msg==0.49.0
fastapi==0.112.0
ffmpy==0.4.0
filelock==3.15.4
flasgger==0.9.5
Flask==2.0.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.2
flask-pyjwt==1.0.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
flask-restx==1.3.0
Flask-SocketIO==5.1.1
flatbuffers==24.3.25
fonttools==4.53.1
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.6.0
future==1.0.0
fuzzywuzzy==0.18.0
gast==0.5.5
gdown==5.2.0
geoip2==4.8.0
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.5.0
google-pasta==0.2.0
googletrans==3.1.0a0
gradio==3.0.13
greenlet==3.0.3
grpcio==1.64.1
gTTS==2.2.3
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
h5py==3.11.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.23.4
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
ifaddr==0.2.0
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==7.2.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
IPy==1.1
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.4.0
kiwisolver==1.4.5
lark==1.1.9
Levenshtein==0.25.1
libclang==18.1.1
linkify-it-py==2.0.3
llvmlite==0.43.0
lxml==5.3.0
maltiverse==1.2.0
mando==0.6.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
maxminddb==2.6.2
mdit-py-plugins==0.4.1
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
ml-dtypes==0.3.2
monotonic==1.6
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
msoffcrypto-tool==5.4.2
multidict==6.0.5
namex==0.0.8
narwhals==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
nudenet==3.4.2
numba==0.60.0
numpy==1.26.4
olefile==0.47
oletools==0.60.2
onnxruntime==1.19.0
openai-whisper==20240930
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.0.10
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.7
packaging==24.1
pandas==1.3.0
paramiko==3.4.0
parso==0.8.4
passlib==1.7.4
pcodedmp==1.2.6
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.3.0
platformdirs==4.3.6
proglog==0.1.10
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==6.0.0
publicsuffixlist==1.0.2.20240903
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
pyee==11.1.0
pygal==3.0.5
pygame==2.6.0
PyGithub==2.3.0
Pygments==2.18.0
PyJWT==2.8.0
pylibsrtp==0.10.0
pymongo==3.12.0
Pympler==1.1
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyotp==2.9.0
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
pypng==0.20220715.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqtgraph==0.13.7
pyreadline3==3.4.1
PySocks==1.7.1
pystyle==2.9
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==0.21.1
python-engineio==4.9.1
python-Levenshtein==0.25.1
python-multipart==0.0.9
python-socketio==5.11.3
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.2.0
qrcode==7.4.2
radon==5.1.0
rapidfuzz==3.9.7
ratelimit==2.2.1
red-black-tree-mod==1.20
referencing==0.35.1
regex==2024.5.15
reportlab==4.2.2
requests==2.32.3
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.20.0
RTFDE==0.1.2
safetensors==0.4.3
scikit-learn==0.24.1
scipy==1.10.1
seaborn==0.13.2
semver==3.0.2
sentencepiece==0.2.0
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
socketIO-client==0.7.2
sounddevice==0.5.0
soupsieve==2.5
SpeechRecognition==3.10.4
sqlmap==1.8.8
stack-data==0.6.3
starlette==0.37.2
streamlit==1.38.0
streamlit-webrtc==0.47.7
sympy==1.12.1
tbb==2021.13.0
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.44.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uc-micro-py==1.0.3
ultralytics==8.2.101
ultralytics-thop==2.0.8
urllib3==1.26.20
uvicorn==0.30.5
validators==0.33.0
watchdog==4.0.2
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==2.3.6
win_unicode_console==0.5
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
